+++ src/FsBunny/Consumer.fs
@@ -5,6 +5,7 @@ module FsBunny.Consumer
 open RabbitMQ.Client
 open RabbitMQ.Client.Events
 open System.Collections.Generic
+open System
 
 let buildConsumer autoAck assember withBinding = 
     let consumerRef = ref Option<QueueingBasicConsumer>.None
@@ -26,20 +27,23 @@ let buildConsumer autoAck assember withBinding =
 
     { new Consumer<'T> with
         member x.Get timeout =
-            let rec consume remaining fraction =
-                match withConsumer (fun consumer -> consumer.Queue.Dequeue fraction) with
-                | true, evt -> 
-                    try 
-                        Some { msg = assember(evt.RoutingKey,evt.BasicProperties.Headers,evt.Body)
-                               id = evt.DeliveryTag }
-                    with ex -> // dead letter policy should be used to handle permanently rejected/undeliverable: http://www.rabbitmq.com/dlx.html
-                        withConsumer (fun consumer -> consumer.Model.BasicNack(evt.DeliveryTag, false, false))
-                        None
-                | false, _ when remaining > 0 -> consume (remaining-fraction) fraction
-                | _ -> None
-            consume (int timeout*1000) (int timeout * 10)
+            match withConsumer (fun consumer -> consumer.Queue.Dequeue (int timeout*1000)) with
+            | true, evt -> 
+                try 
+                    Some { msg = assember(evt.RoutingKey,evt.BasicProperties.Headers,evt.Body)
+                           id = evt.DeliveryTag }
+                with ex -> // dead letter policy should be used to handle permanently rejected/undeliverable: http://www.rabbitmq.com/dlx.html
+                    withConsumer (fun consumer -> consumer.Model.BasicNack(evt.DeliveryTag, false, false))
+                    None
+            | _ -> None
         member x.Ack id = withConsumer <| fun consumer -> consumer.Model.BasicAck(id, false)
         member x.Nack id = withConsumer <| fun consumer -> consumer.Model.BasicNack(id, false, true)
+        member x.Dispose() =
+            match !consumerRef with
+            | Some v -> 
+               v.Model.Dispose()
+               consumerRef := None
+            | _ -> ()
     }
 
 /// Convert auto-acking consumer into ISeq<_>